wrong libusb C header inclusion
libusb.h include path assumes that the file lies in /libusb/,
which might not be true for all systems (on Linux systems
the subdirectory is libusb-1.0). It would be better to just
include only libusb.h and set the sub-directory as include
path via pkg-config or something else.

Error was:

.../src/configuration_impl.hpp:7:27: fatal error:
   libusb/libusb.h: No such file or directory

Related-to-issue: #4
Signed-off-by: Stephan Linz <linz@li-pro.net>