fix char to wchar_t copy failure (type cast error)
Following code fragment implies that sizeof(wchar_t) is equal to 2:

    unsigned char str[128];
    std::wstring wstr;
    wstr.assign((const wchar_t*)str, sizeof(str));

For most modern 64-bit systems and for gcc and clang that is not
true regardless of whether the compiler targets 32-bit or 64-bit
systems. So the implicit assumption wchar_t is 16-bit is wrong
and not portable.

The solution here is to copy byte by byte regardless of bit width.

For further information see:

    http://stackoverflow.com/a/21240783
    http://stackoverflow.com/a/7211215

Signed-off-by: Stephan Linz <linz@li-pro.net>